==> notebooks/0-download_data.ipynb <==
# Download data from cBioPortal 

download link: https://cbioportal-datahub.s3.amazonaws.com/brca_metabric.tar.gz

## 前提条件

カレントディレクトリが/notebooks/であること（breast_cancer_analyticsディレクトリから'jupyter lab or notebookコマンドによってjupyter kernelを立ち上げたこと'）

from os.path import exists, join, dirname
import shutil
from pathlib import Path
import tarfile

import requests
import urllib.error
import urllib.request
import pandas as pd

import config
import functions

### データディレクトリの作成

functions.make_dir(config.RAW_DIR)
functions.make_dir(config.EXTERNAL_DIR)
functions.make_dir(config.INTERIM_DIR)
functions.make_dir(config.PROCESSED_DIR)

### geneデータを含む全データのダウンロード・解凍

# file_name: cBioPortalからダウンロードするzipファイル
file_name = config.RAW_BRCA_METABRIC_TG
try:
    # fileが存在しなければcBioPortalから取得
    if not exists(file_name):
        urlData = requests.get(config.URL_cBioPortal).content
        with open(file_name, mode="wb") as f:
            f.write(urlData)
except urllib.error.URLError as e:
    print(e)

# 解凍先の保存dir
functions.make_dir(config.RAW_BRCA_METABRIC_DIR)
with tarfile.open(config.RAW_BRCA_METABRIC_TG, "r:gz") as t:
    t.extractall(path=config.RAW_DIR)
