==> notebooks/1.1-EDA-metabric.ipynb <==
# EDA metabric data (including clinical patient and sample data)

clinical patient data と clinical sample data の分析  
どちらも患者データに属する. 


import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from config import *
import sweetviz as sv
import pandas_profiling as pdp
import dtale

# データ全体

データ概要の表示  
MB-XXXXとMTS-TXXXX両方を含めた全体での欠損値などを見ていく

df_metabric=pd.read_table(DATA_DIR+'/brca_metabric_clinical_data.tsv')
df_metabric.head()

df_metabric.info()

dtale.show(df_metabric)

df_metabric.shape

# 欠損値の個数
df_metabric.isnull().sum()

# データ分割

IDをMB-XXXXとMTS-TXXXXに分けて、データの密度の高いMB-XXXXについてデータの分布や欠損値を見ていく  
加えてCancer-Type毎にもデータを分け、以降の分析、モデル構築に役立てる

## ID分割

MB-XXXXとMTS-TXXXXでデータを分割し、保存もする

df_metabric_MB=df_metabric[df_metabric['Patient ID'].str.contains('MB')]
df_metabric_MTST=df_metabric[df_metabric['Patient ID'].str.contains('MTS-T')]

df_metabric_MB.describe()

df_metabric_MTST.describe()

df_metabric_MB.isnull().sum()

df_metabric_MB.nunique()==1

他の列と重複しているcolumnと内容が唯一のcolumnをdrop(df_metabric_MB_modified)

df_metabric_MB.columns

# 削除する列
# 理由：一意のみの情報、インデックス、他の列と同意

drop_columns=[
    'Study ID',
    'Patient ID',
    'Number of Samples Per Patient',
    'Sample Type',
    'Sex',
    'Oncotree Code'
]


df_metabric_MB_modified=df_metabric_MB.T.drop_duplicates().T.drop(drop_columns, axis=1)
df_metabric_MB_modified.shape

df_metabric_MB_modified.columns

df_metabric_MB_modified.dtypes

df_metabric_MB_modified.nunique()

dtale.show(df_metabric_MB_modified)

df_metabric_MB_modified.head()

df_metabric_MB_modified.to_csv('tmp_df.csv', index=False)
