==> notebooks/3.0-preprocess_merge_df.ipynb <==
import os

import numpy as np
import pandas as pd

import config

# 最大表示列数の指定（ここでは50列を指定）
pd.set_option("display.max_columns", 50)
pd.set_option("display.max_rows", 50)


def make_dir(dir_name: str):
    if not os.path.exists(dir_name):
        os.makedirs(dir_name)

# basic process

基本的なプロセス  
2つのデータを読み込み、PATIENT_IDカラムをキーとして結合する  
その後、MB\~とMTX-T\~でPATIENT_IDが別れているので、分割し、各データを保存する

df_patient = pd.read_table(
    config.RAW_BRCA_METABRIC_DIR + "/data_clinical_patient.txt", header=4
)
df_sample = pd.read_table(
    config.RAW_BRCA_METABRIC_DIR + "/data_clinical_sample.txt", header=4
)

df_patient.shape, df_sample.shape

# 重複カラム確認
df_tmp = pd.concat([df_patient, df_sample], axis=1)
display(df_tmp.shape)
df_tmp.T.duplicated()

df_patientとdf_sampleをconcatした場合、PATIENT_ID*2とSAMPLE_IDは内容が完全に同じ

# df_patientとdf_sampleを結合
df_merged = pd.merge(df_patient, df_sample)
df_merged.info()

# 　手を加える前のデータフレームの形状
df_merged.shape, df_merged.isnull().sum()

# データを大きく2つに分割できるので、ここで分割
df_MB = df_merged[df_merged["PATIENT_ID"].str.contains("MB")]
df_MTST = df_merged[df_merged["PATIENT_ID"].str.contains("MTS-T")]
df_MB.shape, df_MTST.shape

# save
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_merged.to_pickle(config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.0-df_merged.pkl")
df_MB.to_pickle(config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.0-df_MB.pkl")
df_MTST.to_pickle(config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.0-df_MTST.pkl")
