==> notebooks/3.1-preprocess_handle_null.ipynb <==
import os

import numpy as np
import pandas as pd

import config

# 最大表示列数の指定（ここでは50列を指定）
pd.set_option("display.max_columns", 50)
pd.set_option("display.max_rows", 50)


def make_dir(dir_name: str):
    if not os.path.exists(dir_name):
        os.makedirs(dir_name)

# PATIENT_IDにMBを含むデータのみ含むデータフレームを読み込み
df_MB = pd.read_pickle(config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.0-df_MB.pkl")
df_MB.shape

df_MB.isnull().sum()

# drop all null data case

### df｜df_MB_dropped  
null値を含む行を削除することで、null値に対応する.  

df_MB_dropped = df_MB.dropna(how="any")
display(df_MB_dropped.shape, df_MB_dropped.isnull().sum().sum())
# save data
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_MB_dropped.to_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.1-df_MB_dropped.pkl"
)

# fill null by mean, median value (only float) case

### df｜df_MB_filled_by_mean, df_MB_filled_by_median  
null値を平均値で埋める。  
ただし、カテゴリ変数など数値以外のデータは平均値を取ることができないので、nullのまま（文字列に変換）にする。

注意｜OS_STATUSとOS_MONTH、RFS_STATUSとRFS_MONTH、VITAL_STATUSのnull値を埋めてしまうと、誤回答のデータを作成してしまうことになるので、あらかじめ除いておく。  
outcomeのnullを除いたdf｜df_MB_dropped_outcome  
除く対象｜場合分けして保存するのが億劫なので、今回は全部削除する。

# null値を含むかの確認表示（nullか着目しているcolumnのみ）
display(
    df_MB[
        df_MB["OS_STATUS"].isnull()
        | df_MB["OS_MONTHS"].isnull()
        | df_MB["RFS_STATUS"].isnull()
        | df_MB["RFS_MONTHS"].isnull()
        | df_MB["VITAL_STATUS"].isnull()
    ][["OS_STATUS", "OS_MONTHS", "RFS_STATUS", "RFS_MONTHS", "VITAL_STATUS"]]
)
# dfのサイズ確認（df_MB, df_MBから上記のdfを取り除いたdf）
display(
    df_MB.shape,
    df_MB[
        ~(
            df_MB["OS_STATUS"].isnull()
            | df_MB["OS_MONTHS"].isnull()
            | df_MB["RFS_STATUS"].isnull()
            | df_MB["RFS_MONTHS"].isnull()
            | df_MB["VITAL_STATUS"].isnull()
        )
    ].shape,
)
# df生成
df_MB_dropped_outcome = df_MB[
    ~(
        df_MB["OS_STATUS"].isnull()
        | df_MB["OS_MONTHS"].isnull()
        | df_MB["RFS_STATUS"].isnull()
        | df_MB["RFS_MONTHS"].isnull()
        | df_MB["VITAL_STATUS"].isnull()
    )
]

df_MB_filled_by_mean = df_MB_dropped_outcome.fillna(
    df_MB_dropped_outcome.mean(numeric_only=True)
)
display(df_MB_filled_by_mean.shape, df_MB_filled_by_mean.isnull().sum())
# save data
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_MB_filled_by_mean.to_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.1-df_MB_filled_by_mean.pkl"
)

df_MB_filled_by_median = df_MB_dropped_outcome.fillna(
    df_MB_dropped_outcome.median(numeric_only=True)
)
display(df_MB_filled_by_median.shape, df_MB_filled_by_median.isnull().sum())
# save data
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_MB_filled_by_median.to_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.1-df_MB_filled_by_median.pkl"
)
