==> notebooks/3.4-preprocess_drop_columns.ipynb <==
import os
import numpy as np
import pandas as pd

from IPython.display import HTML
from sklearn.preprocessing import OneHotEncoder, LabelEncoder

import config

# 最大表示列数の指定（ここでは50列を指定）
pd.set_option("display.max_columns", 50)
pd.set_option("display.max_rows", 50)


def make_dir(dir_name: str):
    if not os.path.exists(dir_name):
        os.makedirs(dir_name)

df_MB_OS_status_in_5years = pd.read_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/3.3-df_MB_OS_status_in_5years.pkl"
)
df_MB_OS_status_in_5years.info()

target_columns = [
    # patient
    "OS_MONTHS",
    "RFS_MONTHS",
    "OS_STATUS",
    "RFS_STATUS",
    "VITAL_STATUS",
]

meanless_columns = [
    # patient
    "PATIENT_ID",
    "SEX",
    # sample
    "SAMPLE_ID",
    "SAMPLE_TYPE",
]

# 不要なdfを削除する

### df| df_MB_OS_status_in_5years_dropped
5年後の予後の2値分類のために、不要な特徴量を削除する。  
削除するcolumnsのリスト
- target_columns
- meanless_columns

target = "target"
drop_columns = target_columns + meanless_columns
df_MB_OS_status_in_5years_dropped = df_MB_OS_status_in_5years.drop(drop_columns, axis=1)
df_MB_OS_status_in_5years_dropped.shape

df_MB_OS_status_in_5years_dropped.info()

# save
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_MB_OS_status_in_5years_dropped.to_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR
    + "/3.4-df_MB_OS_status_in_5years_dropped.pkl"
)
