==> notebooks/4.2-preprocess_encoding_columns.ipynb <==
import os
import numpy as np
import pandas as pd

from sklearn.preprocessing import OneHotEncoder, LabelEncoder

import config

# 最大表示列数の指定（ここでは50列を指定）
pd.set_option("display.max_columns", 50)
pd.set_option("display.max_rows", 50)


def make_dir(dir_name: str):
    if not os.path.exists(dir_name):
        os.makedirs(dir_name)

# 型のエンコーディング
## df｜df_MB_dtype_encoded
順序尺度特徴量とboolean特徴量について、データを扱いやすい形に変換する。

df_MB_dtype_encoded = pd.read_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/4.1-df_MB_dropped.pkl"
)
df_MB_dtype_encoded.info()

target_columns = [
    # patient
    "OS_MONTHS",
    "RFS_MONTHS",
    "OS_STATUS",
    "RFS_STATUS",
    "VITAL_STATUS",
]

int_columns = [
    # patient
    "LYMPH_NODES_EXAMINED_POSITIVE",
    "OS_MONTHS",
    "RFS_MONTHS",
]

float_columns = [
    # patient
    "NPI",
    "AGE_AT_DIAGNOSIS",
    # sample
    "TUMOR_SIZE",
    "TMB_NONSYNONYMOUS",
]

str_columns = []

# 質的変数（順序尺度）
qualitative_ordinal_columns = [
    # patient
    "CELLULARITY",
    "ER_IHC",
    "HER2_SNP6",
    "INFERRED_MENOPAUSAL_STATE",
    # sample
    "ER_STATUS",
    "HER2_STATUS",
    "GRADE",
    "PR_STATUS",
    "TUMOR_STAGE",
]
# 質的変数（名義尺度）
qualitative_name_columns = [
    # patient
    "COHORT",
    "INTCLUST",
    "CLAUDIN_SUBTYPE",
    "THREEGENE",
    "HISTOLOGICAL_SUBTYPE",
    "BREAST_SURGERY",
    "LATERALITY",
    "VITAL_STATUS",
    # sample
    "CANCER_TYPE",
    "CANCER_TYPE_DETAILED",
    "ONCOTREE_CODE",
]

bool_columns = [
    # patient
    "CHEMOTHERAPY",
    "HORMONE_THERAPY",
    "RADIO_THERAPY",
    "OS_STATUS",
    "RFS_STATUS",
]

meanless_columns = [
    # patient
    "PATIENT_ID",
    "SEX",
    # sample
    #'PATIENT_ID',
    "SAMPLE_ID",
    "SAMPLE_TYPE",
]

print(
    df_MB_dtype_encoded.shape[1],
    len(int_columns)
    + len(float_columns)
    + len(ｓｔｒ_columns)
    + len(qualitative_ordinal_columns)
    + len(qualitative_name_columns)
    + len(bool_columns)
    + len(meanless_columns),
)
assert df_MB_dtype_encoded.shape[1] == len(int_columns) + len(float_columns) + len(
    ｓｔｒ_columns
) + len(qualitative_ordinal_columns) + len(qualitative_name_columns) + len(
    bool_columns
) + len(
    meanless_columns
), "lack or too much columns"

# 質的変数（順序尺度）のエンコーディング
qualitative_ordinal_columnsは、順序のあるstrの変数で構成されている。
そこで大小関係に対応するようstrをintに変換する。

df_MB_dtype_encoded[qualitative_ordinal_columns].head()

for column in qualitative_ordinal_columns:
    print(column)
    display(df_MB_dtype_encoded[column].value_counts())
    print("-----" * 10)

df_MB_dtype_encoded["CELLULARITY"].replace(
    {"High": 3, "Moderate": 2, "Low": 1}, inplace=True
)
df_MB_dtype_encoded["ER_IHC"].replace({"Positve": 1, "Negative": -1}, inplace=True)
df_MB_dtype_encoded["HER2_SNP6"].replace(
    {"GAIN": 3, "NEUTRAL": 2, "LOSS": 1, "UNDEF": 0}, inplace=True
)
df_MB_dtype_encoded["INFERRED_MENOPAUSAL_STATE"].replace(
    {"Post": 1, "Pre": -1}, inplace=True
)
df_MB_dtype_encoded["ER_STATUS"].replace({"Positive": 1, "Negative": -1}, inplace=True)
df_MB_dtype_encoded["HER2_STATUS"].replace(
    {"Positive": 1, "Negative": -1}, inplace=True
)
# df_MB_dtype_encoded['GRADEGRADE']
df_MB_dtype_encoded["PR_STATUS"].replace({"Positive": 1, "Negative": -1}, inplace=True)
# df_MB_dtype_encoded['TUMOR_STAGE']

for column in qualitative_ordinal_columns:
    print(column)
    display(df_MB_dtype_encoded[column].value_counts())
    print("-----" * 10)

# boolのエンコーディング
bool_columnsは対象的な2項目から構成されているため、これらの項目をboolean型にする。
Yes、Noのみを対象とする（2項目のみから構成される特徴量でも、Yes、No以外の場合は3項目を考慮してカテゴリ特徴量として考える）。

df_MB_dtype_encoded[bool_columns].astype(bool).info()

for column in bool_columns:
    print(column)
    display(df_MB_dtype_encoded[column].value_counts())
    print("-----" * 10)

df_MB_dtype_encoded["CHEMOTHERAPY"].replace({"YES": True, "NO": False}, inplace=True)
df_MB_dtype_encoded["HORMONE_THERAPY"].replace({"YES": True, "NO": False}, inplace=True)
df_MB_dtype_encoded["RADIO_THERAPY"].replace({"YES": True, "NO": False}, inplace=True)
df_MB_dtype_encoded["OS_STATUS"].replace(
    {"1:DECEASED": True, "0:LIVING": False}, inplace=True
)
df_MB_dtype_encoded["RFS_STATUS"].replace(
    {"1:Recurred": True, "0:Not Recurred": False}, inplace=True
)

for column in bool_columns:
    print(column)
    display(df_MB_dtype_encoded[column].value_counts())
    print("-----" * 10)

df_MB_dtype_encoded[bool_columns].dtypes

# save
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_MB_dtype_encoded.to_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/4.2-df_MB_dtype_encoded.pkl"
)

# Onehot encoding
## df｜df_MB_onehot_encoded(onehot encodingしたdf), df_MB_onehot_concated(onehot encodingしたdfと元のdfの結合)
上記で作成した一部特徴量を修正したdfを元に、onehot encodingを行う。
onehot encoding自体は名義尺度の質的変数カラムであるqualitative_name_columnsのみに対してのみ実施するが、最終的に特徴量の型の修正はモデルへ入力する際に必要になるので、df_MB_dtype_encodedを元にonehot encodingを実施する。    

import os
import numpy as np
import pandas as pd

from sklearn.preprocessing import OneHotEncoder, LabelEncoder

import config

# 最大表示列数の指定（ここでは50列を指定）
pd.set_option("display.max_columns", 50)
pd.set_option("display.max_rows", 50)


def make_dir(dir_name: str):
    if not os.path.exists(dir_name):
        os.makedirs(dir_name)

df_MB_dtype_encoded = pd.read_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/4.2-df_MB_dtype_encoded.pkl"
)
df_MB_dtype_encoded.info()

# VITAL_STATUSは目的変数の生成に必要&リークになりうるので、onehot encodingからは除外する
qualitative_name_columns.remove(
    "VITAL_STATUS"
) if "VITAL_STATUS" in qualitative_name_columns else qualitative_name_columns
qualitative_name_columns

# 多重共線性回避のために、drop='first'
onehot_encoder = OneHotEncoder(drop="first")
onehot_encoding_columns = qualitative_name_columns
onehot_encoder.fit(df_MB_dtype_encoded[onehot_encoding_columns])
display(onehot_encoder.categories_)
display(onehot_encoder.get_feature_names_out(onehot_encoding_columns))
# enc.transform(df_MB_onehot_encoded[encoding_columns]).toarray()
df_MB_onehot_encoded = pd.DataFrame(
    onehot_encoder.transform(df_MB_dtype_encoded[onehot_encoding_columns]).toarray(),
    columns=onehot_encoder.get_feature_names_out(onehot_encoding_columns),
)
df_MB_onehot_encoded.shape

# 元のdfとonehotしたdfの結合
dfの形などを調査

2つのdfのshapeを確認

df_MB_dtype_encoded.shape, df_MB_onehot_encoded.shape

# 単に結合しただけでは形が異なる
pd.concat([df_MB_dtype_encoded, df_MB_onehot_encoded], axis=1).shape

# concatがうまく行かない理由はindexの順番。
# 元のdfのnullをdropしたため。

# なのでindexをresetしてからconcatする

# onehotで生成した元のカラムの削除
df_MB_onehot_concated = pd.concat(
    [
        df_MB_dtype_encoded.reset_index(drop=True),
        df_MB_onehot_encoded.reset_index(drop=True),
    ],
    axis=1,
).drop(onehot_encoding_columns, axis=1)
display(df_MB_onehot_concated.shape)

display(df_MB_onehot_concated.info())

# save
make_dir(config.INTERIM_PICKLE_PREPROCESSED_DIR)
df_MB_onehot_concated.to_pickle(
    config.INTERIM_PICKLE_PREPROCESSED_DIR + "/4.2-df_MB_onehot_concated.pkl"
)
